requirements.txt
asgiref==3.7.2
cloudinary==1.38.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-mailjet==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
gunicorn==21.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
